C++ Client: doxygen comments, cleanups, and refactoring
1. Put doxygen comments on many methods
2. Remove some chatty console messages
3. Move TableMaker here from the "tests" repo so that anyone can use it
   if they want
