+++ cpp-client/.gitignore
@@ -1,2 +1,3 @@
+*~
 cmake-build-debug
 cmake-build-release
