+++ cpp-client/deephaven/client/include/public/deephaven/client/highlevel/types.h
@@ -33,10 +33,18 @@ public:
   static constexpr const int64_t MAX_LONG = std::numeric_limits<int64_t>::max();
 };
 
+/**
+ * The Deephaven DBDateTime type. Records nanoseconds relative to the epoch (January 1, 1970) UTC.
+ * Times before the epoch can be represented with negative nanosecond values.
+ */
 class DBDateTime {
 public:
-  // Converts nanos-since-UTC-epoch to DBDateTime. Deephaven null value sentinel is turned into
-  // DBDateTime(0).
+  /**
+   * Converts nanosseconds-since-UTC-epoch to DBDateTime. The Deephaven null value sentinel is
+   * turned into DBDateTime(0).
+   * @param nanos Nanoseconds since the epoch (January 1, 1970 UTC).
+   * @return The corresponding DBDateTime.
+   */
   static DBDateTime fromNanos(long nanos) {
     if (nanos == DeephavenConstants::NULL_LONG) {
       return DBDateTime(0);
@@ -44,12 +52,48 @@ public:
     return DBDateTime(nanos);
   }
 
+  /**
+   * Default constructor. Sets the DBDateTime equal to the epoch.
+   */
   DBDateTime() = default;
+  /**
+   * Sets the DBDateTime to the specified number of nanoseconds relative to the epoch.
+   * @param nanos Nanoseconds since the epoch (January 1, 1970 UTC).
+   */
   explicit DBDateTime(int64_t nanos) : nanos_(nanos) {}
+  /**
+   * Sets the DBDateTime to the specified date, with a time component of zero.
+   * @param year Year.
+   * @param month Month.
+   * @param day Day.
+   */
   DBDateTime(int year, int month, int day);
+  /**
+   * Sets the DBDateTime to the specified date and time, with a fractional second component of zero.
+   * @param year Year.
+   * @param month Month.
+   * @param day Day.
+   * @param hour Hour.
+   * @param minute Minute.
+   * @param second Second.
+   */
   DBDateTime(int year, int month, int day, int hour, int minute, int second);
+  /**
+   * Sets the DBDateTime to the specified date and time, including fractional seconds expressed
+   * in nanos.
+   * @param year Year.
+   * @param month Month.
+   * @param day Day.
+   * @param hour Hour.
+   * @param minute Minute.
+   * @param second Second.
+   * @param nanos Nanoseconds.
+   */
   DBDateTime(int year, int month, int day, int hour, int minute, int second, long nanos);
 
+  /*
+   * The DBDateTime as expressed in nanoseconds since the epoch. Can be negative.
+   */
   int64_t nanos() const { return nanos_; }
 
   void streamIrisRepresentation(std::ostream &result) const;
