+++ cpp-client/doc/.gitignore
@@ -1,0 +1,2 @@
+doxygenoutput/
+sphinxoutput/
