+++ cpp-client/doc/_templates/placeholder.txt
@@ -1,0 +1,2 @@
+This directory is currently empty. This file exists so that the directory
+will be created by git.
