+++ cpp-client/doc/fluent_expressions.rst
@@ -1,0 +1,23 @@
+Fluent classes representing expressions
+=======================================
+
+Leaf classes
+------------
+
+.. doxygenclass:: deephaven::client::highlevel::NumericExpression
+   :members:
+
+.. doxygenclass:: deephaven::client::highlevel::StringExpression
+   :members:
+
+.. doxygenclass:: deephaven::client::highlevel::DateTimeExpression
+   :members:
+
+.. doxygenclass:: deephaven::client::highlevel::BooleanExpression
+   :members:
+
+Intermediate classes
+--------------------
+
+.. doxygenclass:: deephaven::client::highlevel::Expression
+   :members:
