+++ cpp-client/doc/index.rst
@@ -1,0 +1,13 @@
+Deephaven C++ Client API Reference
+==================================
+
+.. toctree::
+   :maxdepth: 2
+
+   main
+   tablehandles
+   fluent
+   types
+   utility
+
+You may also want to consult the :ref:`Index <genindex>`.
