+++ cpp-client/doc/types.rst
@@ -1,0 +1,18 @@
+Other Deephaven Types
+=====================
+
+A variety of other types that appear in the system.
+
+Types used for sorting tables
+-----------------------------
+
+.. doxygenclass:: deephaven::client::highlevel::SortPair
+   :members:
+
+.. doxygenenum:: deephaven::client::highlevel::SortDirection
+
+Types used for manipulating dates/times
+---------------------------------------
+
+.. doxygenclass:: deephaven::client::highlevel::DBDateTime
+   :members:
