+++ cpp-client/doc/utility.rst
@@ -1,0 +1,10 @@
+Utility classes
+===============
+
+These are various utility and helper classes available in the system.
+
+.. doxygenclass:: deephaven::client::utility::TableMaker
+   :members:
+
+.. doxygenclass:: deephaven::client::highlevel::FlightWrapper
+   :members:
