+++ cpp-client/tests/test_util.h
@@ -1,3 +1,6 @@
+/*
+ * Copyright (c) 2016-2020 Deephaven Data Labs and Patent Pending
+ */
 #pragma once
 
 #include <string>
@@ -17,6 +20,7 @@
 #include <arrow/util/key_value_metadata.h>
 #include "deephaven/client/highlevel/client.h"
 #include "deephaven/client/utility/utility.h"
+#include "deephaven/client/utility/table_maker.h"
 
 namespace deephaven {
 namespace client {
@@ -63,178 +67,6 @@ private:
   std::vector<int64_t> volume_;
 };
 
-namespace internal {
-class TypeConverter {
-public:
-  template<typename T>
-  static TypeConverter createNew(const std::vector<T> &values);
-
-  TypeConverter(std::shared_ptr<arrow::DataType> dataType, std::string deephavenType,
-      std::shared_ptr<arrow::Array> column);
-  ~TypeConverter();
-
-  const std::shared_ptr<arrow::DataType> &dataType() const { return dataType_; }
-  std::shared_ptr<arrow::DataType> &dataType() { return dataType_; }
-
-  const std::string &deephavenType() const { return deephavenType_; }
-  std::string &deephavenType() { return deephavenType_; }
-
-  const std::shared_ptr<arrow::Array> &column() const { return column_; }
-  std::shared_ptr<arrow::Array> &column() { return column_; }
-
-private:
-  template<typename T>
-  static const T *tryGetContainedValue(const T *value, bool *valid) {
-    *valid = true;
-    return value;
-  }
-
-  template<typename T>
-  static const T *tryGetContainedValue(const std::optional<T> *value, bool *valid) {
-    if (!value->has_value()) {
-      *valid = false;
-      return nullptr;
-    }
-    *valid = true;
-    return &**value;
-  }
-
-  std::shared_ptr<arrow::DataType> dataType_;
-  std::string deephavenType_;
-  std::shared_ptr<arrow::Array> column_;
-};
-
-template<typename T>
-struct TypeConverterTraits {
-  // The below assert fires when this class is instantiated; i.e. when none of the specializations
-  // match. It needs to be written this way (with "is_same<T,T>") because for technical reasons it
-  // needso be dependent on T, even if degenerately so.
-  static_assert(!std::is_same<T, T>::value, "TableMaker doesn't know how to work with this type");
-};
-
-template<>
-struct TypeConverterTraits<bool> {
-  typedef arrow::BooleanType arrowType_t;
-  typedef arrow::BooleanBuilder arrowBuilder_t;
-  static const char *deephavenTypeName;
-};
-
-template<>
-struct TypeConverterTraits<int8_t> {
-  typedef arrow::Int8Type arrowType_t;
-  typedef arrow::Int8Builder arrowBuilder_t;
-  static const char *deephavenTypeName;
-};
-
-template<>
-struct TypeConverterTraits<int16_t> {
-  typedef arrow::Int16Type arrowType_t;
-  typedef arrow::Int16Builder arrowBuilder_t;
-  static const char *deephavenTypeName;
-};
-
-template<>
-struct TypeConverterTraits<int32_t> {
-  typedef arrow::Int32Type arrowType_t;
-  typedef arrow::Int32Builder arrowBuilder_t;
-  static const char *deephavenTypeName;
-};
-
-template<>
-struct TypeConverterTraits<int64_t> {
-  typedef arrow::Int64Type arrowType_t;
-  typedef arrow::Int64Builder arrowBuilder_t;
-  static const char *deephavenTypeName;
-};
-
-template<>
-struct TypeConverterTraits<float> {
-  typedef arrow::FloatType arrowType_t;
-  typedef arrow::FloatBuilder arrowBuilder_t;
-  static const char *deephavenTypeName;
-};
-
-template<>
-struct TypeConverterTraits<double> {
-  typedef arrow::DoubleType arrowType_t;
-  typedef arrow::DoubleBuilder arrowBuilder_t;
-  static const char *deephavenTypeName;
-};
-
-template<>
-struct TypeConverterTraits<std::string> {
-  typedef arrow::StringType arrowType_t;
-  typedef arrow::StringBuilder arrowBuilder_t;
-  static const char *deephavenTypeName;
-};
-
-template<typename T>
-struct TypeConverterTraits<std::optional<T>> {
-  typedef TypeConverterTraits<T> inner_t;
-  typedef typename inner_t::arrowType_t arrowType_t;
-  typedef typename inner_t::arrowBuilder_t arrowBuilder_t;
-  static const char *deephavenTypeName;
-};
-
-template<typename T>
-const char *TypeConverterTraits<std::optional<T>>::deephavenTypeName =
-    TypeConverterTraits<T>::deephavenTypeName;
-
-template<typename T>
-TypeConverter TypeConverter::createNew(const std::vector<T> &values) {
-  using deephaven::client::utility::flight::statusOrDie;
-  using deephaven::client::utility::stringf;
-
-  typedef TypeConverterTraits<T> traits_t;
-
-  auto dataType = std::make_shared<typename traits_t::arrowType_t>();
-
-  typename traits_t::arrowBuilder_t builder;
-  for (const auto &value : values) {
-    bool valid;
-    const auto *containedValue = tryGetContainedValue(&value, &valid);
-    if (valid) {
-      statusOrDie(builder.Append(*containedValue), "builder.AppendValue");
-    } else {
-      statusOrDie(builder.AppendNull(), "builder.AppendNull");
-    }
-  }
-  auto builderRes = builder.Finish();
-  if (!builderRes.ok()) {
-    auto message = stringf("Error building array of type %o: %o", traits_t::deephavenTypeName,
-        builderRes.status().ToString());
-  }
-  auto array = builderRes.ValueOrDie();
-  return TypeConverter(std::move(dataType), traits_t::deephavenTypeName, std::move(array));
-}
-}  // namespace internal
-
-class TableWizard {
-  typedef deephaven::client::highlevel::TableHandleManager TableHandleManager;
-  typedef deephaven::client::highlevel::TableHandle TableHandle;
-public:
-  TableWizard();
-  ~TableWizard();
-
-  template<typename T>
-  void addColumn(std::string name, const std::vector<T> &values);
-
-  TableHandle makeTable(const TableHandleManager &manager, std::string tableName);
-
-private:
-  void finishAddColumn(std::string name, internal::TypeConverter info);
-
-  arrow::SchemaBuilder schemaBuilder_;
-  int64_t numRows_ = 0;
-  std::vector<std::shared_ptr<arrow::Array>> columns_;
-};
-
-template<typename T>
-void TableWizard::addColumn(std::string name, const std::vector<T> &values) {
-  auto info = internal::TypeConverter::createNew(values);
-  finishAddColumn(std::move(name), std::move(info));
-}
-
 class TableMakerForTests {
   typedef deephaven::client::highlevel::Client Client;
   typedef deephaven::client::highlevel::TableHandleManager TableHandleManager;
@@ -251,17 +83,16 @@ public:
   const ColumnDataForTests &columnData() { return columnData_; }
 
 public:
-  TableMakerForTests(Client &&client, TableHandle &&testTable, std::string &&testTableName,
-      ColumnNamesForTests &&columnNames, ColumnDataForTests &&columnData) : client_(std::move(client)),
-      testTable_(std::move(testTable)), testTableName_(std::move(testTableName)),
-      columnNames_(std::move(columnNames)), columnData_(std::move(columnData)) {}
+  TableMakerForTests(Client &&client, TableHandle &&testTable, ColumnNamesForTests &&columnNames,
+      ColumnDataForTests &&columnData) : client_(std::move(client)),
+      testTable_(std::move(testTable)), columnNames_(std::move(columnNames)),
+      columnData_(std::move(columnData)) {}
 
   const Client &client() const { return client_; }
 
 private:
   Client client_;
   TableHandle testTable_;
-  std::string testTableName_;
   ColumnNamesForTests columnNames_;
   ColumnDataForTests columnData_;
 };
@@ -277,7 +108,7 @@ inline void compareTableRecurse(int depth, const std::shared_ptr<arrow::Table> &
 template<typename T, typename... Args>
 void compareTableRecurse(int depth, const std::shared_ptr<arrow::Table> &table,
     const std::string &columnName, const std::vector<T> &data, Args &&... rest) {
-  TypeConverter zi = TypeConverter::createNew(data);
+  auto zi = deephaven::client::utility::internal::TypeConverter::createNew(data);
   const auto &dataAsArrow = zi.column();
   compareTableHelper(depth, table, columnName, dataAsArrow);
   compareTableRecurse(depth + 1, table, std::forward<Args>(rest)...);
